move components to workflow invocation folder
