+++ client/src/components/WorkflowInvocationState/JobStep.vue
@@ -36,7 +36,7 @@
 <script>
 import Vue from "vue";
 import BootstrapVue from "bootstrap-vue";
-import { JobProvider } from "components/History/providers/DatasetProvider";
+import { JobProvider } from "./providers";
 import JobInformation from "components/JobInformation/JobInformation";
 import JobParameters from "components/JobParameters/JobParameters";
 import UtcDate from "components/UtcDate";
